feat: add animation
